+++ library/dash/src/main/java/com/google/android/exoplayer2/source/dash/offline/DashDownloadAction.java
@@ -21,9 +21,7 @@ import com.google.android.exoplayer2.offline.DownloadAction;
 import com.google.android.exoplayer2.offline.DownloaderConstructorHelper;
 import com.google.android.exoplayer2.offline.SegmentDownloadAction;
 import com.google.android.exoplayer2.offline.StreamKey;
-import java.io.DataInputStream;
-import java.io.DataOutputStream;
-import java.io.IOException;
+import java.util.Collections;
 import java.util.List;
 
 /** An action to download or remove downloaded DASH streams. */
@@ -34,15 +32,6 @@ public final class DashDownloadAction extends SegmentDownloadAction {
 
   public static final Deserializer DESERIALIZER =
       new SegmentDownloadActionDeserializer(TYPE, VERSION) {
-
-        @Override
-        protected StreamKey readKey(DataInputStream input) throws IOException {
-          int periodIndex = input.readInt();
-          int groupIndex = input.readInt();
-          int trackIndex = input.readInt();
-          return new StreamKey(periodIndex, groupIndex, trackIndex);
-        }
-
         @Override
         protected DownloadAction createDownloadAction(
             Uri uri, boolean isRemoveAction, byte[] data, List<StreamKey> keys) {
@@ -51,27 +40,45 @@ public final class DashDownloadAction extends SegmentDownloadAction {
       };
 
   /**
+   * Creates a DASH download action.
+   *
+   * @param uri The URI of the media to be downloaded.
+   * @param data Optional custom data for this action. If {@code null} an empty array will be used.
+   * @param keys Keys of tracks to be downloaded. If empty, all tracks will be downloaded.
+   */
+  public static DashDownloadAction createDownloadAction(
+      Uri uri, @Nullable byte[] data, List<StreamKey> keys) {
+    return new DashDownloadAction(uri, /* isRemoveAction= */ false, data, keys);
+  }
+
+  /**
+   * Creates a DASH remove action.
+   *
+   * @param uri The URI of the media to be removed.
+   * @param data Optional custom data for this action. If {@code null} an empty array will be used.
+   */
+  public static DashDownloadAction createRemoveAction(Uri uri, @Nullable byte[] data) {
+    return new DashDownloadAction(uri, /* isRemoveAction= */ true, data, Collections.emptyList());
+  }
+
+  /**
    * @param uri The DASH manifest URI.
    * @param isRemoveAction Whether the data will be removed. If {@code false} it will be downloaded.
    * @param data Optional custom data for this action.
    * @param keys Keys of representations to be downloaded. If empty, all representations are
    *     downloaded. If {@code removeAction} is true, {@code keys} must be empty.
+   * @deprecated Use {@link #createDownloadAction(Uri, byte[], List)} or {@link
+   *     #createRemoveAction(Uri, byte[])}.
    */
+  @Deprecated
   public DashDownloadAction(
       Uri uri, boolean isRemoveAction, @Nullable byte[] data, List<StreamKey> keys) {
     super(TYPE, VERSION, uri, isRemoveAction, data, keys);
   }
 
   @Override
-  protected DashDownloader createDownloader(DownloaderConstructorHelper constructorHelper) {
+  public DashDownloader createDownloader(DownloaderConstructorHelper constructorHelper) {
     return new DashDownloader(uri, keys, constructorHelper);
-  }
-
-  @Override
-  protected void writeKey(DataOutputStream output, StreamKey key) throws IOException {
-    output.writeInt(key.periodIndex);
-    output.writeInt(key.groupIndex);
-    output.writeInt(key.trackIndex);
   }
 
 }